refactor: restructure module structure
