+++ src/editor/buffer/line.rs
@@ -1,68 +1,6 @@
 use core::fmt;
 
-use unicode_segmentation::UnicodeSegmentation;
-use unicode_width::UnicodeWidthStr;
-
-#[derive(Copy, Clone)]
-enum GraphemeWidth {
-    Half,
-    Full,
-}
-
-impl GraphemeWidth {
-    pub fn from_usize(w: usize) -> Self {
-        match w {
-            0 | 1 => Self::Half,
-            2 => Self::Full,
-            _ => panic!("Invalid grapheme width"),
-        }
-    }
-    pub fn to_usize(&self) -> usize {
-        match self {
-            Self::Half => 1,
-            Self::Full => 2,
-        }
-    }
-}
-
-fn calc_tab_width(current_pos: usize) -> usize {
-    let tabstop = 8;
-    (current_pos / tabstop + 1) * tabstop - current_pos
-}
-
-#[derive(Clone)]
-pub struct Grapheme {
-    string: String,
-    width: GraphemeWidth,
-}
-
-impl Grapheme {
-    pub fn is_tab(&self) -> bool {
-        self.string
-            .chars()
-            .next()
-            .expect("contents of grapheme should not be empty")
-            == '\t'
-    }
-    fn get_width_at_current_pos(&self, current_pos: usize) -> usize {
-        if self.is_tab() {
-            calc_tab_width(current_pos)
-        } else {
-            self.width.to_usize()
-        }
-    }
-}
-
-fn str_to_graphemes(s: &str) -> Vec<Grapheme> {
-    let graphemes = s
-        .graphemes(true)
-        .map(|s| Grapheme {
-            string: String::from(s),
-            width: GraphemeWidth::from_usize(s.width_cjk()),
-        })
-        .collect::<Vec<_>>();
-    graphemes
-}
+use super::grapheme::{str_to_graphemes, Grapheme};
 
 pub struct Line {
     graphemes: Vec<Grapheme>,
@@ -93,7 +31,7 @@ impl Line {
                 }
                 // add fully visible grapheme
                 else {
-                    result.push_str(&grapheme.string);
+                    result.push_str(&grapheme.to_string());
                 }
             }
             current_pos = next_pos;
@@ -116,7 +54,7 @@ impl Line {
             if i == idx {
                 result.push(c);
             }
-            result.push_str(&grapheme.string);
+            result.push_str(&grapheme.to_string());
         }
         if idx >= self.len() {
             result.push(c);
@@ -128,7 +66,7 @@ impl Line {
 
         for (i, grapheme) in self.graphemes.iter().enumerate() {
             if i != idx {
-                result.push_str(&grapheme.string);
+                result.push_str(&grapheme.to_string());
             }
         }
         self.graphemes = str_to_graphemes(&result);
@@ -148,7 +86,11 @@ impl Line {
 
 impl fmt::Display for Line {
     fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        let result: String = self.graphemes.iter().map(|g| g.string.clone()).collect();
+        let result: String = self
+            .graphemes
+            .iter()
+            .map(|g| g.to_string().clone())
+            .collect();
         write!(formatter, "{result}")
     }
 }
