--- src/editor/window/location.rs
@@ -1,5 +1,0 @@
-#[derive(Copy, Clone, Default)]
-pub struct Location {
-    pub x: usize,
-    pub y: usize,
-}
